3.5.0-dev-1 Saving data to a database now works correctly. Added 24/7 mode to player and command /247. Command /help now displays prefixes. Renaming of audioplayer components.
