+++ package.json
@@ -29,7 +29,7 @@
     "@distube/spotify": "^2.0.2",
     "@distube/youtube": "^1.0.2",
     "@distube/yt-dlp": "^2.0.1",
-    "@distube/ytdl-core": "^4.13.7",
+    "@distube/ytdl-core": "^4.14.1",
     "@distube/ytsr": "^2.0.4",
     "@eslint/js": "^9.8.0",
     "cross-env": "7.0.3",
@@ -55,17 +55,16 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/node": "^20.14.13",
+    "@types/node": "^20.14.14",
     "@types/node-cron": "^3.0.11",
     "@types/node-os-utils": "^1.3.4",
-    "@types/puppeteer": "^7.0.4",
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "eslint": "^9.8.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
-    "globals": "^15.8.0",
+    "globals": "^15.9.0",
     "prettier": "^3.3.3",
     "typescript": "^5.5.4"
   }
