+++ src/DiscordTypes.ts
@@ -1,12 +1,12 @@
 import { Client, ClientEvents, Collection } from 'discord.js';
 import { ICommand, ICommandGroup } from './CommandTypes.js';
-import { AudioPlayerCore } from './audioplayer/AudioPlayerCore.js';
+import { AudioPlayersManager } from './audioplayer/AudioPlayersManager.js';
 
 declare module 'discord.js' {
   export interface Client {
     commands: Collection<string, ICommand>;
     commandsGroups: Collection<string, ICommandGroup>;
-    audioPlayer: AudioPlayerCore;
+    audioPlayer: AudioPlayersManager;
   }
 }
