+++ src/audioplayer/MessagePlayerButtonsHandler.ts
@@ -197,7 +197,7 @@ export class MessagePlayerButtonsHandler {
             break;
 
           // case ButtonIDs.downloadSong: {
-          //     const song = this.client.audioplayer.distube.getQueue(ButtonInteraction.guild as Guild)?.songs[0]
+          //     const song = this.client.audioplayer.distube.getQueue(ButtonInteraction.guild as guild)?.songs[0]
           //
           //     if (!song) {
           //         await ButtonInteraction.reply({embeds: [generateErrorEmbed(i18next.t("audioplayer:download_song_error"))]})
