+++ src/locales/en/audioplayer.json
@@ -27,7 +27,8 @@
   "player_embed_song_length": "Song duration",
   "player_embed_next_song": "Next song",
   "player_embed_next_song_empty": "Empty",
-  "player_embed_state_waiting": "Waiting next song",
+  "player_embed_state_waiting_leave_on_empty_true": "Waiting next song, if song is not provided, bot leave channel in 20 seconds",
+  "player_embed_state_waiting_leave_on_empty_false": "Waiting next song, 24/7 mode is enabled, so bot not leave channel",
   "player_embed_state_pause": "Pause",
   "player_embed_state_playing": "Playing",
   "player_embed_state_loading": "Please, wait...",
