+++ src/locales/ru/audioplayer.json
@@ -27,7 +27,8 @@
   "player_embed_song_length": "Длительность песни",
   "player_embed_next_song": "Следующая песня",
   "player_embed_next_song_empty": "Пусто",
-  "player_embed_state_waiting": "Жду следующую песню",
+  "player_embed_state_waiting_leave_on_empty_true": "Жду следующую песню, если песни не будет, бот покинет канал через 20 секунд",
+  "player_embed_state_waiting_leave_on_empty_false": "Жду следующую песню, 24/7 режим включён, бот не будет покидать голосовой канал",
   "player_embed_state_pause": "Пауза",
   "player_embed_state_playing": "Играет",
   "player_embed_state_loading": "Пожалуйста, подождите...",
