+++ src/locales/ru/commands.json
@@ -71,5 +71,8 @@
   "lyrics_desc": "Ищет текст песни",
   "lyrics_arg_query": "Название песни",
   "lyrics_embed_text_not_correct": "Текст может быть неверным",
-  "lyrics_embed_lyrics_not_found": "Текст песни не найден"
+  "lyrics_embed_lyrics_not_found": "Текст песни не найден",
+  "247_desc": "Переключает 24/7 режим. Если включено, бот остаётся в канале, даже когда закончились песни.",
+  "247_enabled": "Режим 24/7 включён",
+  "247_disabled": "Режим 24/7 отключён"
 }
