+++ src/main.ts
@@ -5,7 +5,7 @@ loggerSend(`Starting bot on version ${process.env.npm_package_version}`);
 import { Client, Partials } from 'discord.js';
 import { loggerError, loggerSend } from './utilities/logger.js';
 import { loginBot } from './utilities/loginBot.js';
-import { AudioPlayerCore } from './audioplayer/AudioPlayerCore.js';
+import { AudioPlayersManager } from './audioplayer/AudioPlayersManager.js';
 import loadLocale from './locales/Locale.js';
 
 await loadLocale();
@@ -21,7 +21,7 @@ client.rest.on('rateLimited', (args) => {
   loggerError(`Client encountered a rate limit: ${JSON.stringify(args)}`);
 });
 
-new AudioPlayerCore(client);
+new AudioPlayersManager(client);
 
 await handlersLoad(client);
